Refactor and clean up components
