+++ src/Components/AuthForm/AuthForm.js
@@ -3,7 +3,7 @@ import { useAuth0 } from '@auth0/auth0-react';
 import './AuthForm.css'
 
 const AuthForm = () => {
-  const { loginWithRedirect, isLoading, logout } = useAuth0();
+  const { loginWithRedirect, isLoading } = useAuth0();
   return (
     <section className='loginLayout'>
       {isLoading && <p>Loading...</p>}
@@ -11,7 +11,7 @@ const AuthForm = () => {
         <h2>Learn all about Weather Vine Here:</h2>
         <p>Weather Vine is an educational app for consumers to connect more deeply with the wine they enjoy. Explore wines from a region and see how the climate has influenced the very wine one drinks!</p>
         <button onClick={() => loginWithRedirect()}>Log In</button>
-        
+
       </article>
     </section>
   )
