+++ src/Components/Header/Header.css
@@ -18,6 +18,10 @@
   margin-left: 20vw;
 }
 
+.hidden {
+  display: none;
+}
+
 @media only screen and (min-width: 1500px){
   .headerLayout {
     height: 12vh;
